+++ resources/views/admin/roles/index.blade.php
@@ -85,7 +85,8 @@
 
             </div>
         @endforeach
-    </div>
+    </section>
+
     <div class="modal fade" id="createRoleModal" tabindex="-1">
         <div class="modal-dialog modal-dialog-centered">
             <div class="modal-content">
@@ -112,7 +113,7 @@
 
             </div>
         </div>
-    </section>
+    </div>
 @endsection
 
 <script>